refactor(types): split the _pantr_cpp stub into a package, one file per area
src/pantr/_pantr_cpp.pyi was 1464 hand-written lines, and a dozen tickets
now each have to add a type to it. It becomes a stub package with one
module per binding area -- _basis, _quad, _bezier, _grid, _geometry,
_transform -- so a ticket that ports a type edits its own file plus one
import line. Not one declared signature changes; the 40 symbols are moved
verbatim. mypy over src tests scripts stays clean, at 290 files where it
was 284, which is the six new stub modules.

mypy.ini sets strict = True, so implicit re-export is off and __init__.pyi
has to spell every name in the `X as X` form. Verified by removing one
`as`: mypy reports 7 errors in 2 files.

The split also introduces a hazard, which _backend.py now handles. A
directory of .pyi files with no __init__.py is an importable NAMESPACE
package. Python's finder prefers a real extension whenever one exists, so
a built installation is unaffected -- but with no extension the probe
`import pantr._pantr_cpp` stopped raising and started returning an empty
module, which reported the C++ backend as available on a Numba-only
install. Measured before the fix: tests/parity/test_geometry_aabb.py gave
81 failures where 97 skips belong. A namespace package has no __file__;
the probe now says so, and test_dispatch.py pins it in both directions.
