+++ src/pantr/_pantr_cpp/_geometry.pyi
@@ -1,0 +1,46 @@
+"""Geometry types owned by the C++ core.
+
+Bound by ``cpp/bindings/geometry.cpp``. See ``__init__.pyi`` for what this
+package promises and who has to keep it.
+"""
+
+import numpy as np
+import numpy.typing as npt
+
+class AABB:
+    """Axis-aligned bounding box owned by the C++ core.
+
+    The first type this extension exposes rather than a free function, under the
+    2026-08-27 amendment to ``design/cross_backend_types.md``. Wrapped by
+    :class:`pantr.geometry.AABB`, which is the class a caller holds; this one is
+    reached only through it.
+
+    Attributes:
+        ndim (int): Spatial dimension, ``>= 1``.
+        lo (npt.NDArray[np.float64]): Lower corner, shape ``(ndim,)``, read-only.
+        hi (npt.NDArray[np.float64]): Upper corner, shape ``(ndim,)``, read-only.
+    """
+
+    def __init__(self, lo: npt.NDArray[np.float64], hi: npt.NDArray[np.float64]) -> None: ...
+    @property
+    def ndim(self) -> int: ...
+    @property
+    def lo(self) -> npt.NDArray[np.float64]: ...
+    @property
+    def hi(self) -> npt.NDArray[np.float64]: ...
+    @staticmethod
+    def unbounded(ndim: int) -> AABB: ...
+    @staticmethod
+    def empty(ndim: int) -> AABB: ...
+    @staticmethod
+    def from_bounds(bounds: npt.NDArray[np.float64]) -> AABB: ...
+    def as_bounds(self) -> npt.NDArray[np.float64]: ...
+    def is_empty(self) -> bool: ...
+    def contains_point(self, x: npt.NDArray[np.float64]) -> bool: ...
+    def overlaps(self, other: AABB) -> bool: ...
+    def union(self, other: AABB) -> AABB: ...
+    def intersect(self, other: AABB) -> AABB | None: ...
+    def pad(self, r: float | npt.NDArray[np.float64]) -> AABB: ...
+    def transform(
+        self, matrix: npt.NDArray[np.float64], offset: npt.NDArray[np.float64]
+    ) -> AABB: ...
